fix(shared): break circular dependency in registry/package type schemas
Extract four low-dependency registry schemas (RegistryReasonCodeSchema,
RegistryTrustTierSchema, RegistryDistributionStatusSchema,
RegistryCompatibilityStateSchema) and RegistryInstallEligibilitySnapshotSchema
into a new registry-eligibility.ts file.

Also move CanonicalPackageTypeSchema from package-manifest.ts to enums.ts
and update package-store.ts to import from enums.ts.

The cycle: registry.ts -> package-manifest.ts -> package-resolution.ts ->
package-lifecycle.ts -> registry.ts caused RegistryReleaseSubmissionInputSchema
to be undefined at test-time, failing 10 tests across subcortex-registry,
subcortex-nudges, cli, and web.

All 1828 tests now pass. Discovered during Phase 14 phase-close verification.
