+++ self/shared/src/types/registry.ts
@@ -7,37 +7,27 @@ import {
 } from './package-manifest.js';
 import { PackageDependencySetSchema } from './package-resolution.js';
 
-export const RegistryReasonCodeSchema = z
-  .string()
-  .regex(/^MKT-00[1-9]-[A-Z0-9][A-Z0-9_-]*$/);
-export type RegistryReasonCode = z.infer<typeof RegistryReasonCodeSchema>;
-
-export const RegistryTrustTierSchema = z.enum([
-  'nous_first_party',
-  'verified_maintainer',
-  'community_unverified',
-  'unregistered_external',
-]);
-export type RegistryTrustTier = z.infer<typeof RegistryTrustTierSchema>;
-
-export const RegistryDistributionStatusSchema = z.enum([
-  'active',
-  'hold',
-  'delisted',
-  'blocked',
-]);
-export type RegistryDistributionStatus = z.infer<
-  typeof RegistryDistributionStatusSchema
->;
-
-export const RegistryCompatibilityStateSchema = z.enum([
-  'compatible',
-  'requires_migration',
-  'blocked_incompatible',
-]);
-export type RegistryCompatibilityState = z.infer<
-  typeof RegistryCompatibilityStateSchema
->;
+// Re-export eligibility schemas from registry-eligibility.ts to keep backward compatibility
+// while breaking the circular dependency with package-lifecycle.ts
+export {
+  RegistryReasonCodeSchema,
+  type RegistryReasonCode,
+  RegistryTrustTierSchema,
+  type RegistryTrustTier,
+  RegistryDistributionStatusSchema,
+  type RegistryDistributionStatus,
+  RegistryCompatibilityStateSchema,
+  type RegistryCompatibilityState,
+  RegistryInstallEligibilitySnapshotSchema,
+  type RegistryInstallEligibilitySnapshot,
+} from './registry-eligibility.js';
+import {
+  RegistryReasonCodeSchema,
+  RegistryTrustTierSchema,
+  RegistryDistributionStatusSchema,
+  RegistryCompatibilityStateSchema,
+  RegistryInstallEligibilitySnapshotSchema,
+} from './registry-eligibility.js';
 
 export const RegistryModerationStateSchema = z.enum([
   'flagged_for_review',
@@ -262,25 +252,6 @@ export const RegistryEligibilityRequestSchema = z.object({
 });
 export type RegistryEligibilityRequest = z.infer<
   typeof RegistryEligibilityRequestSchema
->;
-
-export const RegistryInstallEligibilitySnapshotSchema = z.object({
-  project_id: ProjectIdSchema.optional(),
-  package_id: z.string().min(1),
-  release_id: z.string().min(1),
-  package_version: z.string().min(1),
-  trust_tier: RegistryTrustTierSchema,
-  distribution_status: RegistryDistributionStatusSchema,
-  compatibility_state: RegistryCompatibilityStateSchema,
-  metadata_valid: z.boolean(),
-  signer_valid: z.boolean(),
-  requires_principal_override: z.boolean().default(false),
-  block_reason_codes: z.array(RegistryReasonCodeSchema).default([]),
-  evidence_refs: z.array(z.string().min(1)).default([]),
-  evaluated_at: z.string().datetime(),
-});
-export type RegistryInstallEligibilitySnapshot = z.infer<
-  typeof RegistryInstallEligibilitySnapshotSchema
 >;
 
 export const RegistryReleaseSubmissionInputSchema = z.object({
